+++ src/main/java/com/my/challenger/dto/quiz/AnswerValidationResult.java
@@ -5,6 +5,8 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 @NoArgsConstructor
@@ -13,7 +15,7 @@ public class AnswerValidationResult {
     private boolean correct;
     private boolean exactMatch;        // true if Levenshtein/exact matched
     private boolean aiAccepted;        // true if AI accepted as equivalent
-    private double aiConfidence;       // AI confidence score (0 if not used)
+    private BigDecimal aiConfidence;       // AI confidence score (0 if not used)
     private String aiExplanation;      // AI explanation (null if not used)
     private boolean aiUsed;            // whether AI was actually called
 }
